style: DTO data 타입, 생성자 변경

+++ src/main/java/com/widyu/healthcare/dto/reward/RewardDTO.java
@@ -11,7 +11,7 @@ public class RewardDTO {
     @NonNull
     long userIdx;
     String description;
-    private Integer status;
+    private byte status;
     String url;
     long point;
